requirements.txt
selenium==4.39.0
python-dotenv==1.2.1